+++ README.md
@@ -26,7 +26,7 @@ In this project, our design team has already built an HTML/CSS web page for us, 
 
 #### Stretch Goals:
 * Research Javascript `import` statements and Break your classes found in `Tabs.js` into their own files. 
-* Break out your `TabItem` and `TabLink` styling into their own LESS files. Import them into your `index.less` file. 
+[x] Break out your `TabItem` and `TabLink` styling into their own LESS files. Import them into your `index.less` file. 
 * Create a Single `Tabs` component that creates all instances of `TabLink`. It will also hold the currently selected tab. 
   * Add a `deselect` method to each the `TabItem` and `TabLink` classes. This will deselect that specific tab.
   * Remove the code deselecting all `TabLink`s and `TabItem`s, only relying on the deselect method.
