+++ components/Tabs/TabLink.js
@@ -1,0 +1,37 @@
+import TabItem from './TabItem.js'
+
+export default class TabLink {
+  constructor(link) {
+    this.link = link;
+    
+    // Get the custom data attribute on the Link
+    this.data = this.link.dataset.tab;
+    
+    // Using the custom data attribute get the associated Item element
+    this.itemElement = document.querySelector(`.tabs-item[data-tab="${this.data}"]`);
+    
+    // Using the Item element, create a new instance of the TabItem class
+    this.tabItem = new TabItem(this.itemElement);
+    
+    // Add a click event listener on this instance, calling the select method on click
+    this.link.addEventListener('click', () => { this.select() });
+  };
+
+  select() {
+    // Get all of the elements with the tabs-link class
+    const links = document.querySelectorAll('.tabs-link');
+    console.log(links)
+
+    // Using a loop or the forEach method remove the 'tabs-link-selected' class from all of the links
+    Array.from(links).forEach( link => {
+      link.classList.remove('tabs-link-selected');
+    });
+
+    // Add a class named "tabs-link-selected" to this link
+    this.link.classList.add('tabs-link-selected');
+    
+    // Call the select method on the item associated with this link
+    this.tabItem.select();
+    console.log('Selected!');
+  }
+}
