+++ index.html
@@ -6,7 +6,8 @@
 	</head>
 	<body>
 		<script defer="defer" src="./components/Dropdown/Dropdown.js"></script>
-		<script defer="defer" src="./components/Tabs/Tabs.js"></script>
+		<script type="module" defer="defer" src="./components/Tabs/TabLink.js"></script>
+		<script type="module" defer="defer" src="./components/Tabs/TabItem.js"></script>
 		
 		<div class="header-container">
 			<div class="dropdown">
